Set things up to time experiments properly

+++ scripts/aws_client.sh
@@ -35,19 +35,23 @@ git submodule update
 # remove obsolete log files.
 rm -rf *.log
 
-sudo apt-get update
+if [[ ! -f "lock" ]]
+then
+  touch "lock"
+  sudo apt-get update
 
-# Dependencies
-sudo apt-get install -y g++
-sudo apt-get install -y build-essential
-sudo apt-get install -y valgrind
+  # Dependencies
+  sudo apt-get install -y g++
+  sudo apt-get install -y build-essential
+  sudo apt-get install -y valgrind
 
-# Bazel
-sudo apt install curl gnupg
-curl -fsSL https://bazel.build/bazel-release.pub.gpg | gpg --dearmor > bazel.gpg
-sudo mv bazel.gpg /etc/apt/trusted.gpg.d/
-echo "deb [arch=amd64] https://storage.googleapis.com/bazel-apt stable jdk1.8" | sudo tee /etc/apt/sources.list.d/bazel.list
-sudo apt-get update && sudo apt-get install -y bazel
+  # Bazel
+  sudo apt install curl gnupg
+  curl -fsSL https://bazel.build/bazel-release.pub.gpg | gpg --dearmor > bazel.gpg
+  sudo mv bazel.gpg /etc/apt/trusted.gpg.d/
+  echo "deb [arch=amd64] https://storage.googleapis.com/bazel-apt stable jdk1.8" | sudo tee /etc/apt/sources.list.d/bazel.list
+  sudo apt-get update && sudo apt-get install -y bazel
+fi
 
 # Compile with bazel
 bazel build ...
